Modifled to upload parquet files in a directory

+++ s3-parquet.py
@@ -37,12 +37,72 @@ def connect_s3():
     )
     return fs
 
-def delete_bucket():
-    file_system = s3fs.S3FileSystem()
-    file_system.rm('s3://my-bucket/foo.txt')  # single file
-    files = ['s3://my-bucket/bar.txt', 's3://my-bucket/baz.txt']
-    file_system.rm(files)  # several files
+#
+def delete_parquet_s3(_bucket, _filename):
 
+    fs = connect_s3()
+    s3_filepath = "s3://" + _bucket
+    s3_filepaths = [path for path in fs.ls(s3_filepath)
+       if path.endswith('.parquet')]
+
+    print("[ Check parquet file location ]")
+    print(s3_filepaths)
+    # print("")
+
+    # exit(1)
+
+    # s3_filepath = _bucket + "/" + _filename + ".parquet"
+    s3_filepath = _bucket + "/" + _filename
+    # pf = pq.ParquetDataset(s3_filepath,filesystem=fs)
+
+    # print("[ List the content of parquet file ]")
+    # print(pf.read().to_pandas())
+
+    fs.rm(s3_filepaths)  # single file
+
+    exit(1)
+
+    if os.path.isdir(_filename):
+        for _fn in os.listdir(_filename):
+            with fs.open(_bucket_name+'/'+'/'+_filename+'/'+_fn, 'wb') as f:
+                df.to_parquet(f)
+            # s3_filepath = _bucket_name+'/'+_fn
+            # s3_filepath = _bucket_name+'/'+_fn+'.parquet'
+            s3_filepath = _bucket_name+'/'+_fn
+            pq.write_to_dataset(Table.from_pandas(df),s3_filepath,filesystem=fs,use_dictionary=True,compression="snappy", version="2.4",)
+            read_parquet_s3( _bucket_name, _fn )
+ 
+
+    file_system.rm('s3://my-bucket/foo.txt')  # single file
+
+    # files = ['s3://my-bucket/bar.txt', 's3://my-bucket/baz.txt']
+    # file_system.rm(files)  # several files
+
+
+    df = pd.DataFrame({'data': np.random.random((1000,))})
+
+    if os.path.isdir(_filename):
+        for _fn in os.listdir(_filename):
+            with fs.open(_bucket_name+'/'+'/'+_filename+'/'+_fn, 'wb') as f:
+                df.to_parquet(f)
+            # s3_filepath = _bucket_name+'/'+_fn
+            # s3_filepath = _bucket_name+'/'+_fn+'.parquet'
+            s3_filepath = _bucket_name+'/'+_fn
+            pq.write_to_dataset(Table.from_pandas(df),s3_filepath,filesystem=fs,use_dictionary=True,compression="snappy", version="2.4",)
+            read_parquet_s3( _bucket_name, _fn )
+    elif os.path.isfile(_filename):
+        with fs.open(_bucket_name+'/'+'/'+_filename, 'wb') as f:
+            df.to_parquet(f)
+        # s3_filepath = _bucket_name+'/'+_filename+'.parquet'
+        s3_filepath = _bucket_name+'/'+_filename
+        pq.write_to_dataset(Table.from_pandas(df),s3_filepath,filesystem=fs,use_dictionary=True,compression="snappy", version="2.4",)
+        read_parquet_s3( _bucket_name, _filename )
+    else:
+       usage()
+
+
+
+#
 def test_exists(self, dataset, mocked_csvs_in_s3):
     assert PartitionedDataSet(mocked_csvs_in_s3, dataset).exists()
 
@@ -52,6 +112,7 @@ def test_exists(self, dataset, mocked_csvs_in_s3):
     s3fs.S3FileSystem().mkdir(empty_folder)
     assert not PartitionedDataSet(empty_folder, dataset).exists()
 
+#
 def test_protocol_usage(self, filepath, instance_type):
     data_set = CSVDataSet(filepath=filepath)
     assert isinstance(data_set._fs, instance_type)
@@ -62,28 +123,19 @@ def test_protocol_usage(self, filepath, instance_type):
     else:
         path = data_set._fs._strip_protocol(filepath)
 
-def test_remote(self):
-    endpoint_url = "http://obs.eu-de.otc.t-systems.com"
-    s3 = s3fs.S3FileSystem(anon=True, client_kwargs=dict(endpoint_url=endpoint_url))
-    s3_store = s3fs.S3Map(root="cyanoalert/cyanoalert-olci-lswe-l2c-v1.zarr", s3=s3, check=False)
-    diagnostic_store = DiagnosticStore(s3_store, logging_observer(log_path='remote-cube.log'))
-    xr.open_zarr(diagnostic_store)
-
-def write_s3():
-    fs = s3fs.S3FileSystem()
-    pa.parquet.write_to_dataset(table, root_path=output_folder, filesystem=fs,
-    compression='snappy', partition_filename_cb=lambda x: filename)
-
+#
 def convert_parquet_to_csv( _local_filepath ):
     df = pd.read_parquet( _local_filepath )
     # print( _local_filepath.split('.')[0] + '.csv' )
     df.to_csv( _local_filepath.split('.')[0] + '.csv' )
     # df.to_csv('filename.csv')
 
+#
 def convert_csv_to_parquet( _local_filepath ):
     df = pd.read_csv( _local_filepath )
     df.to_parquet( _local_filepath.split('.')[0] + '-test-' + '.parquet' )
 
+#
 def read_parquet_local( _local_filepath ):
     pf = pq.ParquetDataset( _local_filepath )
     print("[ Read data ]")
@@ -92,31 +144,36 @@ def read_parquet_local( _local_filepath ):
     # print(pf.fragments)
     # print(pf.schema)
 
+#
 def write_parquet_local( _local_filepath ):
     df = pd.DataFrame({'data': np.random.random((1000,))})
     df.to_parquet( _local_filepath )
 
+#
 def write_parquet_s3( _bucket_name, _filename ):
     fs = connect_s3()
     df = pd.DataFrame({'data': np.random.random((1000,))})
-    # df.to_parquet( _local_filepath )
 
-    with fs.open(_bucket_name+'/'+_filename+'.parquet', 'wb') as f:
-        df.to_parquet(f)
+    if os.path.isdir(_filename):
+        for _fn in os.listdir(_filename):
+            with fs.open(_bucket_name+'/'+'/'+_filename+'/'+_fn, 'wb') as f:
+                df.to_parquet(f)
+            # s3_filepath = _bucket_name+'/'+_fn
+            # s3_filepath = _bucket_name+'/'+_fn+'.parquet'
+            s3_filepath = _bucket_name+'/'+_fn
+            pq.write_to_dataset(Table.from_pandas(df),s3_filepath,filesystem=fs,use_dictionary=True,compression="snappy", version="2.4",)
+            read_parquet_s3( _bucket_name, _fn )
+    elif os.path.isfile(_filename):
+        with fs.open(_bucket_name+'/'+'/'+_filename, 'wb') as f:
+            df.to_parquet(f)
+        # s3_filepath = _bucket_name+'/'+_filename+'.parquet'
+        s3_filepath = _bucket_name+'/'+_filename
+        pq.write_to_dataset(Table.from_pandas(df),s3_filepath,filesystem=fs,use_dictionary=True,compression="snappy", version="2.4",)
+        read_parquet_s3( _bucket_name, _filename )
+    else:
+       usage()
 
-    s3_filepath = _bucket_name+'/'+_filename+'.parquet'
-
-    pq.write_to_dataset(
-        Table.from_pandas(df),
-        s3_filepath,
-        filesystem=fs,
-        use_dictionary=True,
-        compression="snappy",
-        version="2.4",
-    )
-
-    read_parquet_s3( _bucket_name, _filename )
-
+#
 def read_parquet_s3( _bucket_name, _filename ):
     fs = connect_s3()
     s3_filepath = "s3://" + _bucket_name
@@ -127,8 +184,8 @@ def read_parquet_s3( _bucket_name, _filename ):
     print(s3_filepaths)
     print("")
 
-    s3_filepath = _bucket_name + "/" + _filename + ".parquet"
     # s3_filepath = _bucket_name + "/" + _filename + ".parquet"
+    s3_filepath = _bucket_name + "/" + _filename
     pf = pq.ParquetDataset(s3_filepath,filesystem=fs)
 
     print("[ List the content of parquet file ]")
@@ -136,34 +193,45 @@ def read_parquet_s3( _bucket_name, _filename ):
 
     # dataset = pq.ParquetDataset('dataset_name/', use_legacy_dataset=False, filters=[('n_legs','=',4)])
 
+#
 def usage():
   print("Usage: ./s3-parquet.py [ -r < read|write> ] [ -l <local|s3> ] [ -b < bucket_name > ] [ -f < file_name > ] [ -d < directory_name > ]")
   return
 
+#
 def main():
     try:
-        opts, args = getopt.getopt(sys.argv[1:],"c:f:l:b:s:t:i:o:",["usage"])
+        opts, args = getopt.getopt(sys.argv[1:],"c:f:l:b:d:s:t:",["command",
+                                                                      "file",
+                                                                      "location",
+                                                                      "bucket",
+                                                                      "directory",
+                                                                      "source",
+                                                                      "target",
+                                                                      "help",
+                                                                      ])
     except getopt.GetoptError as err:
         print(str(err))
         usage()
         sys.exit(1)
 
+#
     for opt,arg in opts:
-        if (opt == "-c"):
+        if opt in ('-c', '--command'):
             _command=arg
-        elif ( opt == "-l"):
+        elif opt in ('-l', '--location'):
             _location=arg
-        elif ( opt == "-b"):
+        elif opt in ('-b', '--bucket'):
             _bucket=arg
-        elif ( opt == "-f"):
-            _files=arg
-        elif ( opt == "-s"):
-            _source=arg
-        elif ( opt == "-t"):
-            _target=arg
-        elif ( opt == "-d"):
+        elif opt in ('-d', '--directory'):
             _directory=arg
-        elif ( opt == "-h") or ( opt == "--help"):
+        elif opt in ('-f', '--file'):
+            _file=arg
+        elif opt in ('-s', '--source'):
+            _source=arg
+        elif opt in ('-t', '--target'):
+            _target=arg
+        elif opt in ('-h', '--help'):
             usage()
         else:
             usage()
@@ -171,22 +239,25 @@ def main():
     # Call functions according to options given
     _sample_data=os.getenv("SAMPLE_DATA")
 
+#
     if ( _command  == "read" ) and ( _location == "local" ):
         read_parquet_local( _sample_data )
     elif ( _command  == "write" ) and ( _location == "local" ):
         write_parquet_local( _sample_data )
-    elif ( _command == "convert" ) and ( _files != "" ) and ( _source == "parquet" ) and  ( _target  == "csv" ):
-        convert_parquet_to_csv( _files )
-    elif ( _command == "convert" ) and ( _files != "" ) and ( _source == "csv" ) and  ( _target  == "parquet" ):
-        convert_csv_to_parquet( _files )
-    elif ( _command == "read" ) and ( _location == "s3" ) and ( _files != "" ):
-        read_parquet_s3( _bucket, _files )
-    elif ( _command == "write" ) and ( _location == "s3" ) and ( _directory != "" ):
+    elif ( _command == "convert" ) and ( _file != "" ) and ( _source == "parquet" ) and  ( _target  == "csv" ):
+        convert_parquet_to_csv( _file )
+    elif ( _command == "convert" ) and ( _file != "" ) and ( _source == "csv" ) and  ( _target  == "parquet" ):
+        convert_csv_to_parquet( _file )
+    elif ( _command == "write" ) and ( _location == "s3" ) and ( _directory is not None ):
         write_parquet_s3( _bucket, _directory )
-    elif ( _command == "write" ) and ( _location == "s3" ) and ( _files != "" ):
-        write_parquet_s3( _bucket, _files )
+    elif ( _command == "read" ) and ( _location == "s3" ) and ( _file is not None ):
+        read_parquet_s3( _bucket, _file )
+    elif ( _command == "delete" ) and ( _location == "s3" ) and ( _file is not None ):
+        delete_parquet_s3( _bucket, _file )
+    elif ( _command == "write" ) and ( _location == "s3" ) and ( _file is not None ):
+        write_parquet_s3( _bucket, _file )
     else:
-        usage
+        usage()
 
     return
